fix: update fetchCredentialStatusVC to use documentLoader
