+++ packages/w3c-context/src/index.ts
@@ -1,1 +1,2 @@
-export * from './lib';
+export * from './lib/index';
+export * from './lib/types';
