+++ packages/w3c-context/src/lib/types.ts
@@ -6,3 +6,12 @@ export type ContextDocument = {
 
 // Define the type for the DID document
 export type Document = ContextDocument | DIDDocument;
+
+export type DocumentLoaderObject = {
+  contextUrl: string | null;
+  document: Document;
+  documentUrl: string;
+};
+// Define a type for the context loader function
+
+export type DocumentLoader = (url: string) => Promise<DocumentLoaderObject>;
